end project with js
