+++ package.json
@@ -11,7 +11,7 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.2.0",
     "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.27.2",
+    "axios": "^1.7.2",
     "clsx": "^1.1.1",
     "easymde": "^2.16.1",
     "prettier": "^2.6.2",
@@ -21,7 +21,7 @@
     "react-markdown": "^8.0.3",
     "react-redux": "^8.0.2",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^3.0.1",
     "react-simplemde-editor": "^5.0.2",
     "sass": "^1.52.1",
     "web-vitals": "^2.1.4"
