+++ src/App.js
@@ -13,7 +13,7 @@ function App() {
 
     React.useEffect(() => {
         dispatch(fetchAuthMe());
-    }, []);
+    });
 
     return (
         <>
