+++ src/axios.js
@@ -1,7 +1,7 @@
 import axios from "axios";
 
 const instance = axios.create({
-    baseURL: "https://backend-blog-gules.vercel.app/",
+    baseURL: "http://localhost:4444/",
 });
 
 instance.interceptors.request.use((config) => {
