+++ src/components/SideBlock/SideBlock.module.scss
@@ -1,7 +1,7 @@
 .root {
-  margin-bottom: 20px;
+    margin-bottom: 20px;
 }
 
 .title {
-  padding: 15px 15px 0 15px;
+    padding: 15px 15px 0 15px;
 }
