+++ src/components/SideBlock/index.jsx
@@ -4,12 +4,12 @@ import Typography from "@mui/material/Typography";
 import Paper from "@mui/material/Paper";
 
 export const SideBlock = ({ title, children }) => {
-  return (
-    <Paper classes={{ root: styles.root }}>
-      <Typography variant="h6" classes={{ root: styles.title }}>
-        {title}
-      </Typography>
-      {children}
-    </Paper>
-  );
+    return (
+        <Paper classes={{ root: styles.root }}>
+            <Typography variant="h6" classes={{ root: styles.title }}>
+                {title}
+            </Typography>
+            {children}
+        </Paper>
+    );
 };
