+++ src/components/TagsSideInfo/index.jsx
@@ -8,7 +8,7 @@ import TagIcon from "@mui/icons-material/Tag";
 import ListItemText from "@mui/material/ListItemText";
 import Skeleton from "@mui/material/Skeleton";
 
-import { SideBlock } from "./SideBlock";
+import { SideBlock } from "../SideBlock";
 
 export const TagsBlock = ({ items, isLoading = true }) => {
     return (
