+++ src/components/index.js
@@ -1,7 +1,7 @@
-export * from "./TagsBlock";
-export * from "./CommentsBlock";
-export * from "./Post";
 export * from "./AddComment";
-export * from "./SideBlock";
-export * from "./UserInfo";
+export * from "./CommentsBlock";
 export * from "./Header";
+export * from "./Post";
+export * from "./SideBlock";
+export * from "./TagsSideInfo";
+export * from "./UserInfo";
