+++ src/pages/SearchTags/index.jsx
@@ -3,21 +3,13 @@ import React from "react";
 import { useSelector } from "react-redux";
 import { useParams } from "react-router-dom";
 import { Post } from "../../components";
-// import { fetchPosts } from "../../redux/slices/posts";
 import axios from "../../axios";
 
 export const SearchTags = () => {
-    // const dispatch = useDispatch();
     const { name } = useParams();
-    const { posts } = useSelector((state) => state.posts);
     const userData = useSelector((state) => state.auth.data);
-    const isPostsLoading = posts.status === "loading";
     const [filteredPosts, setFilteredPosts] = React.useState();
     const [isLoading, setIsLoading] = React.useState(true);
-
-    // const filteredPosts = posts.items.filter((obj, index) => {
-    //     return obj.tags.includes(name) ? obj : "";
-    // });
 
     React.useEffect(() => {
         axios
