+++ src/redux/slices/posts.js
@@ -11,11 +11,6 @@ export const fetchPosts = createAsyncThunk(
     }
 );
 
-export const fetchTags = createAsyncThunk("posts/fetchTags", async () => {
-    const { data } = await axios.get("/tags");
-    return data;
-});
-
 export const fetchRemovePost = createAsyncThunk(
     "posts/fetchRemovePost",
     async (id) => {
@@ -25,10 +20,6 @@ export const fetchRemovePost = createAsyncThunk(
 
 const initialState = {
     posts: {
-        items: [],
-        status: "loading",
-    },
-    tags: {
         items: [],
         status: "loading",
     },
@@ -51,18 +42,6 @@ const postsSlice = createSlice({
             .addCase(fetchPosts.rejected, (state) => {
                 state.posts.items = [];
                 state.posts.status = "error";
-            })
-            // Получение тегов
-            .addCase(fetchTags.pending, (state) => {
-                state.tags.status = "loading";
-            })
-            .addCase(fetchTags.fulfilled, (state, action) => {
-                state.tags.items = action.payload;
-                state.tags.status = "loaded";
-            })
-            .addCase(fetchTags.rejected, (state) => {
-                state.tags.items = [];
-                state.tags.status = "error";
             })
             // Удаление статьи
             .addCase(fetchRemovePost.pending, (state, action) => {
