+++ src/theme.js
@@ -5,7 +5,6 @@ export const theme = createTheme({
         "none",
         "0px 1px 3px rgba(0, 0, 0, 0.2), 0px 1px 1px rgba(0, 0, 0, 0.14), 0px 2px 1px -1px rgba(0, 0, 0, 0.12)",
         "0px 1px 5px rgba(0, 0, 0, 0.2), 0px 2px 2px rgba(0, 0, 0, 0.14), 0px 3px 1px -2px rgba(0, 0, 0, 0.12)",
-        // Добавьте больше теней по необходимости
     ],
     palette: {
         primary: {
